add tests for noaa_hycom
